requirements.txt
Jinja2
SQLAlchemy==1.3.6
Flask
gunicorn
python-dotenv